display redeploy if user is going to redeploy contract to their account
